requirements.txt
langchain_chroma==0.1.4
langchain_community==0.3.7
langchain_core==0.3.18
langchain_ollama==0.2.0
langchain_text_splitters==0.3.2
python-dotenv==1.0.1
